Rename createHandlers to AirviewMockServer
