--- packages/airview-msw-handlers/src/features/create-handlers/index.js
@@ -1,1 +1,0 @@
-export { AirviewMockServer } from "./create-handlers";
