+++ packages/airview-msw-handlers/src/features/index.js
@@ -1,1 +1,1 @@
-export * from "./create-handlers";
+export * from "./airview-mock-server";
